// src/Types/ProductTypes.ts
export interface IProduct {
  cric_member_product_id: number;
  cric_member_product_name: string;
  cric_member_product_price: number;
  cric_member_special_product_price: number;
  cric_member_product_category: string;
  cric_member_product_url: string;
  cric_member_product_description: string;
  cric_member_product_photo: string[];
  user_cric_member_product_status: string;
  user_cric_member_company_name: string;
}